Refactor/Full-App (#15)
* Added SaveChanges inside the Commit Method in UnitOfWork

* Moved some debendencies from infra-layer to app and api layers

* Moved helpers from shared-layer to app and domain layers

* Moved Router.cs from shared to api layer, moved aspnet url and location services from application to api layer, removed shared layer

* Updated JwtSettings for readability, no logic inside infrastructure layer

+++ SchoolProject.Application/Features/ApplicationRole/Commands/DeleteRole/DeleteRoleByIdHandler.cs
@@ -1,5 +1,6 @@
 using AutoMapper;
 using MediatR;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Localization;
 using SchoolProject.Application.Bases;
 using SchoolProject.Application.Interfaces.IdentityServices;
@@ -28,6 +29,9 @@ public class DeleteRoleByIdHandler : ResponseHandler, IRequestHandler<DeleteRole
     public async Task<Response<string>> Handle(DeleteRoleByIdCommand request, CancellationToken cancellationToken)
     {
         var role = await _roleManager.GetByIdAsync(request.Id);
+        if (await _roleManager.IsRoleInUseAsync(role!.Name!))
+            return BadRequest<string>(_localizer[SharedResourceKeys.RoleHasUsers]);
+
         await _roleManager.DeleteAsync(role!);
         return Deleted<string>();
     }
